Create README.md
Putting on weight is as difficult as losing weight, particularly when one has a fast metabolism or when one cannot consume enough calories. Although most people are more interested in losing weight, gaining weight in a healthy way takes as much planning and discipline as losing weight. The trick is to gain muscle or not fat in a combination of healthy foods, strength workouts and habits. This is the whole, do-not-fail plan of gaining weight within a month.